V1.0 uncleaned code but deployed *Fixat hur vi hämtar "timeNow" till att använda unixTime istället. TO-DO: ta bort sekundräknaren helt och hållet då den inte längre behövs.
*ändrat getData flaggan så vi skickar datarequest tills vi faktiskt fått
ett svar från servern, eventuellt förläng periodiciteten så vi inte
bombar servern.

* ändrat hur lång tid "soluppgången" ska ta och ändrat lite i
temperatur-paletten samt vind-paletten

*Försökt minska antalet droppar vid regn/snö

*Plus annat småfixande.

Detta är versionen som skickades till skåne och är stress-skriven för
att hinna färdigt

+++ src/hlpFunc.h
@@ -14,9 +14,9 @@ struct weatherDataS{
 };
 
 struct sunTimesS{
-  int timeToRise;
-  int timeToSet;
-  int timeNow;
+  unsigned long timeToRise;
+  unsigned long timeToSet;
+  unsigned long timeNow;
 };
 
 struct mappedDataS{
